backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "liboqs-python @ file:///${PROJECT_ROOT}/libs/liboqs-python",
    "pytest>=8.4.1",
    "requests>=2.32.4",
]
